Move QtDims to _qt
