+++ napari/_qt/qtviewer/main.py
@@ -7,7 +7,7 @@ from qtpy.QtWidgets import QWidget, QVBoxLayout, QSplitter, QFileDialog
 from qtpy.QtGui import QCursor, QPixmap
 from vispy.scene import SceneCanvas, PanZoomCamera
 
-from ...components._dims.qtdims import QtDims
+from napari._qt.qtdims import QtDims
 from ...components._layers_list.view import QtLayersList
 from ...resources import resources_dir
 from ...util.theme import template
